V2/research/hole_optimization/requirements.txt
# Core dependencies
jax>=0.4.0
jaxlib>=0.4.0
optax>=0.1.0
numpy>=1.21.0
scipy>=1.8.0
pydantic>=2.0.0
PyYAML>=6.0

# Experiment management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Optional dependencies for specific optimizers
# Uncomment as needed:

# For genetic algorithms
# deap>=1.3.0
# pymoo>=0.6.0

# For integer programming
# pyscipopt>=4.0.0
# pulp>=2.6.0
# ortools>=9.0.0

# For enhanced optimization
# scikit-optimize>=0.9.0
# hyperopt>=0.2.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
ruff>=0.1.0
mypy>=1.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Experiment tracking (optional)
# wandb>=0.15.0
# tensorboard>=2.10.0

# High-performance computing (optional)
# ray>=2.0.0
# dask>=2022.0.0
